polish(web): round 2 — shared inputs, viewer role, token drift, ID truncation
Remaining candidates from the v1.1 UI polish survey:

- Admin audit filters and the teams invite email field now use the
  shared Input component instead of hand-rolled inputs that suppressed
  the global focus ring; audit filter labels gain explicit htmlFor/id
  wiring.
- The viewer org role was styled in ROLE_CONFIG but missing from every
  role select and invite option, despite OrganizationRole.VIEWER being
  a valid backend value; it is now assignable and invitable on the
  teams and organizations pages.
- text-white replaced with the text-sc-on-accent token on accent
  surfaces (button, gradient button, pagination, pills tabs, checkbox,
  empty-state CTA, not-found, reset-password) so a future on-accent
  tweak lands everywhere.
- The three popover/toolbar search inputs (tasks tags, entity types,
  graph nodes) had three different focus treatments; unified on the
  app-standard cyan ring.
- EmptyState/ErrorState/SuccessState moved out of tooltip.tsx into
  empty-state.tsx with re-exports preserving existing import paths.
- Memory source breadcrumb and import-progress chips no longer render
  full raw entity IDs: truncated with title tooltips, and chip hover
  stays cyan instead of flipping to purple.

Survey item 10 (Organizations page duplicating Teams member
management) intentionally deferred - removing a page's functionality
is a product call, not polish.
